+++ src/components/Header/Header.js
@@ -7,6 +7,7 @@ const Header = () => {
 				<NavLink to="/">Overview</NavLink>
 				<NavLink to="/transactions">Transactions</NavLink>
 				<NavLink to="/adddbtransactions">Add Transactions</NavLink>
+				<NavLink to="/dbtransactions">App Transactions</NavLink>
 				<NavLink to="/goals">Goals</NavLink> //*this piece is combined with the
 				Calculations piece*//
 				<NavLink to="/bills">Bills</NavLink>
